requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
dataclasses==0.7
decorator==4.4.2
deprecation==2.1.0
eyeD3==0.9.5
filetype==1.0.7
gmusicapi==13.0.0
gpsoauth==0.4.1
httplib2==0.18.1
idna==2.10
lxml==4.5.1
MechanicalSoup==0.12.0
mutagen==1.44.0
oauth2client==4.1.3
packaging==20.4
proboscis==1.2.6.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.9.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
validictory==1.1.2
